fixed sky pystest
